Move modellib to models, move models.py into models, move Access into its own file. (#2776)
