--- backend/audit/modellib/__init__.py
@@ -1,8 +1,0 @@
-from .submission_event import (  # noqa
-    SubmissionEvent,
-)
-
-# In case we want to iterate through all the models for some reason:
-models = [
-    SubmissionEvent,
-]
