+++ backend/audit/models/__init__.py
@@ -1,0 +1,32 @@
+from .access import Access
+from .models import (
+    ExcelFile,
+    GeneralInformationMixin,
+    LateChangeError,
+    SingleAuditChecklist,
+    SingleAuditChecklistManager,
+    SingleAuditReportFile,
+    User,
+    excel_file_path,
+    generate_sac_report_id,
+    single_audit_report_path,
+)
+from .submission_event import SubmissionEvent
+
+# In case we want to iterate through all the models for some reason:
+_models = [
+    Access,
+    ExcelFile,
+    GeneralInformationMixin,
+    SubmissionEvent,
+    LateChangeError,
+    SingleAuditChecklist,
+    SingleAuditChecklistManager,
+    SingleAuditReportFile,
+    User,
+]
+_functions = [
+    excel_file_path,
+    generate_sac_report_id,
+    single_audit_report_path,
+]
